Take the bidirectional algorithm and the glyph ink from forme
Both were decided in forme (its PRs #10 and #11) and this is the other side of
them.

internal/bidi and cmd/genbidi are deleted: 3,264 lines implementing UAX #9 a
second time, together with the generated Bidi_Class tables, the conformance
harness and the Makefile targets that fetched Unicode's suites. forme/bidi
serves both callers now, and ADR 0007 records why ADR 0006's reasoning stopped
holding — the arguments in it were all about forme's API surface, and each was
answerable by exporting something.

Levels are int rather than uint8, which is forme's representation and carries
one more thing: -1 for a character rule X9 removed, filled in per §5.2 before
a caller ever sees it.

textInk asks the face where the run's own glyphs put ink instead of where the
face reserves room for the deepest tail it has. A four-em ellipsis is three
dots on the baseline; asked of Courier's descent it appears to hang ten times
as far below the line as it does, and an "overflow: hidden" box that ends
under the baseline appears to cut it. Worth four reftests — line-clamp-002,
line-clamp-007, floats-154 and inlines-020 — so the clean-pass baseline rises
from 4434 to 4438.

The dependency is not bumped yet: this builds against a sibling checkout
through a gitignored go.work, and needs a forme release once those two PRs
land.

+++ render/inline.go
@@ -376,7 +376,7 @@ type inlineItem struct {
 	// The flag is separate because zero is a real level, the left-to-right one,
 	// and a box with no content on the line at all has to stay distinguishable
 	// from a box whose content is left-to-right.
-	insetLevel      uint8
+	insetLevel      int
 	insetLevelKnown bool
 	// float is the box of a float met in this run of inline content. It carries
 	// no text of its own: it is a marker saying "a float belongs here", because
@@ -454,7 +454,7 @@ type inlineItem struct {
 	// that needs no reordering, which is what tells the line builder there is
 	// nothing to do.
 	para  *bidiParagraph
-	level uint8
+	level int
 }
 
 // inlineContent lays a box's inline children into lines and returns the height
